% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{Test function that logs a string into the console.}
\usage{
test_function(input_str)
}
\arguments{
\item{input_str}{String to be logged into R console.}
}
\description{
Test function that logs a string into the console.
}
